move to components/replay
