+++ packages/app/src/cli/services/function/ui/components/Replay/Replay.test.tsx
@@ -1,9 +1,9 @@
 import {Replay} from './Replay.js'
 import {useFunctionWatcher} from './hooks/useFunctionWatcher.js'
-import {testFunctionExtension, testApp} from '../../../../models/app/app.test-data.js'
-import {ExtensionInstance} from '../../../../models/extensions/extension-instance.js'
-import {FunctionConfigType} from '../../../../models/extensions/specifications/function.js'
-import {FunctionRunData} from '../../replay.js'
+import {testFunctionExtension, testApp} from '../../../../../models/app/app.test-data.js'
+import {ExtensionInstance} from '../../../../../models/extensions/extension-instance.js'
+import {FunctionConfigType} from '../../../../../models/extensions/specifications/function.js'
+import {FunctionRunData} from '../../../replay.js'
 import {AbortController} from '@shopify/cli-kit/node/abort'
 import React from 'react'
 import {beforeAll, describe, expect, test, vi} from 'vitest'
