+++ packages/app/src/cli/services/function/ui/components/Replay/hooks/useFunctionWatcher.test.tsx
@@ -1,7 +1,7 @@
 import {useFunctionWatcher} from './useFunctionWatcher.js'
-import {FunctionRunData} from '../../../replay.js'
-import {testApp, testFunctionExtension} from '../../../../../models/app/app.test-data.js'
-import {setupExtensionWatcher} from '../../../../dev/extension/bundler.js'
+import {FunctionRunData} from '../../../../replay.js'
+import {testApp, testFunctionExtension} from '../../../../../../models/app/app.test-data.js'
+import {setupExtensionWatcher} from '../../../../../dev/extension/bundler.js'
 import {Replay} from '../Replay.js'
 import {AbortError} from '@shopify/cli-kit/node/error'
 import {AbortController} from '@shopify/cli-kit/node/abort'
