+++ packages/app/src/cli/services/function/ui/components/Replay/hooks/useFunctionWatcher.ts
@@ -1,8 +1,8 @@
-import {FunctionRunData} from '../../../replay.js'
-import {AppInterface} from '../../../../../models/app/app.js'
-import {FunctionConfigType} from '../../../../../models/extensions/specifications/function.js'
-import {ExtensionInstance} from '../../../../../models/extensions/extension-instance.js'
-import {setupExtensionWatcher} from '../../../../dev/extension/bundler.js'
+import {FunctionRunData} from '../../../../replay.js'
+import {AppInterface} from '../../../../../../models/app/app.js'
+import {FunctionConfigType} from '../../../../../../models/extensions/specifications/function.js'
+import {ExtensionInstance} from '../../../../../../models/extensions/extension-instance.js'
+import {setupExtensionWatcher} from '../../../../../dev/extension/bundler.js'
 import {exec} from '@shopify/cli-kit/node/system'
 import {AbortController} from '@shopify/cli-kit/node/abort'
 import {useEffect, useRef, useState} from 'react'
